Refactor #68 - move package `CompletedScheduleShownRepository`
