+++ app/src/main/java/com/nlab/reminder/internal/feature/schedule/all/di/AllScheduleRepositoryModule.kt
@@ -20,9 +20,9 @@ import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import com.nlab.reminder.BuildConfig
-import com.nlab.reminder.domain.common.schedule.CompletedScheduleShownRepository
+import com.nlab.reminder.domain.common.schedule.visibleconfig.CompletedScheduleShownRepository
 import com.nlab.reminder.domain.feature.schedule.all.AllScheduleScope
-import com.nlab.reminder.internal.common.schedule.impl.LocalCompletedScheduleShownRepository
+import com.nlab.reminder.internal.common.schedule.visibleconfig.impl.LocalCompletedScheduleShownRepository
 import dagger.Module
 import dagger.Provides
 import dagger.Reusable
