+++ app/src/test/java/com/nlab/reminder/domain/feature/schedule/all/AllScheduleGenerator.kt
@@ -18,6 +18,7 @@ package com.nlab.reminder.domain.feature.schedule.all
 
 import com.nlab.reminder.core.effect.SideEffectHandle
 import com.nlab.reminder.domain.common.schedule.*
+import com.nlab.reminder.domain.common.schedule.visibleconfig.CompletedScheduleShownRepository
 import com.nlab.reminder.test.genBoolean
 import kotlinx.coroutines.flow.emptyFlow
 import org.mockito.kotlin.doReturn
